+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/view/ViewBox.kt
@@ -16,19 +16,17 @@
 package app.cash.redwood.layout.view
 
 import android.content.Context
-import android.view.Gravity
 import android.view.View
+import android.view.View.LAYOUT_DIRECTION_LTR
+import android.view.View.MeasureSpec
+import android.view.View.MeasureSpec.makeMeasureSpec
 import android.view.ViewGroup
-import android.view.ViewGroup.LayoutParams.MATCH_PARENT
-import android.view.ViewGroup.LayoutParams.WRAP_CONTENT
-import android.widget.FrameLayout
-import androidx.core.view.updateLayoutParams
 import app.cash.redwood.Modifier
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.layout.modifier.Height
 import app.cash.redwood.layout.modifier.HorizontalAlignment
-import app.cash.redwood.layout.modifier.Margin as MarginModifier
+import app.cash.redwood.layout.modifier.Margin as RedwoodMargin
 import app.cash.redwood.layout.modifier.Size
 import app.cash.redwood.layout.modifier.VerticalAlignment
 import app.cash.redwood.layout.modifier.Width
@@ -36,17 +34,18 @@ import app.cash.redwood.layout.widget.Box
 import app.cash.redwood.ui.Density
 import app.cash.redwood.ui.Margin
 import app.cash.redwood.widget.ViewGroupChildren
+import app.cash.redwood.widget.Widget
 
 internal class ViewBox(
   context: Context,
-) : FrameLayout(context),
+) : ViewGroup(context),
   Box<View> {
   private val density = Density(context.resources)
   private var horizontalAlignment = CrossAxisAlignment.Start
   private var verticalAlignment = CrossAxisAlignment.Start
-  private var width = Constraint.Wrap
-  private var height = Constraint.Wrap
-  private var margin: Margin? = null
+  private var widthConstraint = Constraint.Wrap
+  private var heightConstraint = Constraint.Wrap
+  private var margin: Margin = Margin.Zero
 
   override var modifier: Modifier = Modifier
 
@@ -54,171 +53,269 @@ internal class ViewBox(
 
   override val children = ViewGroupChildren(this)
 
-  override fun generateDefaultLayoutParams(): LayoutParams {
-    return LayoutParams(WRAP_CONTENT, WRAP_CONTENT)
-  }
+  private val measurer = Measurer()
 
   override fun width(width: Constraint) {
-    this.width = width
-    invalidate()
+    this.widthConstraint = width
+    requestLayout()
   }
 
   override fun height(height: Constraint) {
-    this.height = height
-    invalidate()
-  }
-
-  override fun setLayoutParams(params: ViewGroup.LayoutParams?) {
-    if (params != null) {
-      margin?.let { margin ->
-        maybeUpdateLayoutParams(margin, params)
-      }
-    }
-    super.setLayoutParams(params)
+    this.heightConstraint = height
+    requestLayout()
   }
 
   override fun margin(margin: Margin) {
     this.margin = margin
-
-    layoutParams?.let { params ->
-      maybeUpdateLayoutParams(margin, params)
-      // Write instance back out to indicate it has changed.
-      super.setLayoutParams(params)
-    }
-  }
-
-  private fun maybeUpdateLayoutParams(margin: Margin, params: ViewGroup.LayoutParams) {
-    if (params is MarginLayoutParams) {
-      with(density) {
-        params.marginStart = margin.start.toPxInt()
-        params.topMargin = margin.top.toPxInt()
-        params.marginEnd = margin.end.toPxInt()
-        params.bottomMargin = margin.bottom.toPxInt()
-      }
-    }
+    requestLayout()
   }
 
   override fun horizontalAlignment(horizontalAlignment: CrossAxisAlignment) {
     this.horizontalAlignment = horizontalAlignment
-    invalidate()
+    requestLayout()
   }
 
   override fun verticalAlignment(verticalAlignment: CrossAxisAlignment) {
     this.verticalAlignment = verticalAlignment
-    invalidate()
+    requestLayout()
   }
 
-  /** Flush Redwood's modifiers into FrameLayout's LayoutParams before it measures. */
   override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-    val widthMode = if (width == Constraint.Fill) {
-      MeasureSpec.EXACTLY
-    } else {
-      MeasureSpec.getMode(widthMeasureSpec)
-    }
     val widthSize = MeasureSpec.getSize(widthMeasureSpec)
-
-    val heightMode = if (height == Constraint.Fill) {
-      MeasureSpec.EXACTLY
-    } else {
-      MeasureSpec.getMode(heightMeasureSpec)
-    }
+    val widthMode = MeasureSpec.getMode(widthMeasureSpec)
     val heightSize = MeasureSpec.getSize(heightMeasureSpec)
+    val heightMode = MeasureSpec.getMode(heightMeasureSpec)
 
-    for (child in children.widgets) {
-      child.value.updateLayoutParams<LayoutParams> {
-        setFrom(child.modifier)
-      }
+    // Short circuit if this box's dimensions are already fully specified.
+    if (widthMode == MeasureSpec.EXACTLY && heightMode == MeasureSpec.EXACTLY) {
+      setMeasuredDimension(widthSize, heightSize)
+      return
     }
 
-    super.onMeasure(
-      MeasureSpec.makeMeasureSpec(widthSize, widthMode),
-      MeasureSpec.makeMeasureSpec(heightSize, heightMode),
+    measurer.box(
+      layoutDirection = layoutDirection,
+      boxDensity = density,
+      boxHorizontalAlignment = horizontalAlignment,
+      boxVerticalAlignment = verticalAlignment,
+      boxMargin = margin,
+      boxWidth = widthSize,
+      boxWidthUnspecified = widthMode == MeasureSpec.UNSPECIFIED,
+      boxHeight = heightSize,
+      boxHeightUnspecified = heightMode == MeasureSpec.UNSPECIFIED,
     )
+
+    var maxWidth = 0
+    var maxHeight = 0
+    for (widget in children.widgets) {
+      measurer.measure(widget = widget)
+      maxWidth = maxOf(maxWidth, measurer.width + measurer.marginWidth)
+      maxHeight = maxOf(maxHeight, measurer.height + measurer.marginHeight)
+    }
+
+    val resultWidth = when {
+      widthMode == MeasureSpec.EXACTLY -> widthSize
+      widthMode == MeasureSpec.AT_MOST -> maxWidth.coerceAtMost(widthSize)
+      else -> maxWidth
+    }
+    val resultHeight = when {
+      heightMode == MeasureSpec.EXACTLY -> heightSize
+      heightMode == MeasureSpec.AT_MOST -> maxHeight.coerceAtMost(heightSize)
+      else -> maxHeight
+    }
+
+    setMeasuredDimension(resultWidth, resultHeight)
   }
 
-  private fun LayoutParams.setFrom(modifier: Modifier) {
-    var horizontalAlignment = horizontalAlignment
-    var verticalAlignment = verticalAlignment
-    var requestedWidth = Int.MIN_VALUE
-    var requestedHeight = Int.MIN_VALUE
+  override fun onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int) {
+    measurer.box(
+      layoutDirection = layoutDirection,
+      boxDensity = density,
+      boxHorizontalAlignment = horizontalAlignment,
+      boxVerticalAlignment = verticalAlignment,
+      boxMargin = margin,
+      boxWidth = (right - left).coerceAtLeast(0),
+      boxHeight = (bottom - top).coerceAtLeast(0),
+    )
 
-    modifier.forEachScoped { childModifier ->
-      // Check for modifier overrides in the children, otherwise default to the Box's alignment
-      // values.
-      when (childModifier) {
-        is HorizontalAlignment -> {
-          horizontalAlignment = childModifier.alignment
-        }
+    for (widget in children.widgets) {
+      measurer.measure(widget)
+      measurer.layout(widget)
+    }
+  }
+}
 
-        is VerticalAlignment -> {
-          verticalAlignment = childModifier.alignment
-        }
+/**
+ * Measures and lays out one child view at a time.
+ *
+ * This class is mutable and reused to avoid object allocation.
+ */
+private class Measurer {
+  // Inputs from the box.
+  var layoutDirection = LAYOUT_DIRECTION_LTR
+  var boxDensity = Density(1.0)
+  var boxHorizontalAlignment = CrossAxisAlignment.Start
+  var boxVerticalAlignment = CrossAxisAlignment.Start
+  var boxMarginStart = 0
+  var boxMarginEnd = 0
+  var boxMarginTop = 0
+  var boxMarginBottom = 0
+  var boxWidthUnspecified = false
+  var boxHeightUnspecified = false
 
-        is Width -> {
-          requestedWidth = with(density) { childModifier.width.toPxInt() }
-        }
+  // The available space for the child view and its margins.
+  var frameWidth = -1
+  var frameHeight = -1
 
-        is Height -> {
-          requestedHeight = with(density) { childModifier.height.toPxInt() }
-        }
+  // Inputs from the child widget.
+  var horizontalAlignment = CrossAxisAlignment.Start
+  var verticalAlignment = CrossAxisAlignment.Start
+  var marginStart = 0
+  var marginEnd = 0
+  var marginTop = 0
+  var marginBottom = 0
+  var requestedWidth = -1
+  var requestedHeight = -1
 
-        is Size -> {
-          requestedWidth = with(density) { childModifier.width.toPxInt() }
-          requestedHeight = with(density) { childModifier.height.toPxInt() }
-        }
+  // Measurement results.
+  var width = -1
+  var height = -1
 
-        is MarginModifier -> {
-          with(density) {
-            marginStart = childModifier.margin.start.toPxInt()
-            marginEnd = childModifier.margin.end.toPxInt()
-            topMargin = childModifier.margin.top.toPxInt()
-            bottomMargin = childModifier.margin.bottom.toPxInt()
+  val marginWidth: Int
+    get() = marginStart + marginEnd
+  val marginHeight: Int
+    get() = marginTop + marginBottom
+
+  /** Configure the enclosing box. */
+  fun box(
+    layoutDirection: Int,
+    boxDensity: Density,
+    boxHorizontalAlignment: CrossAxisAlignment,
+    boxVerticalAlignment: CrossAxisAlignment,
+    boxMargin: Margin,
+    boxWidth: Int,
+    boxWidthUnspecified: Boolean = false,
+    boxHeight: Int,
+    boxHeightUnspecified: Boolean = false,
+  ) {
+    this.layoutDirection = layoutDirection
+    this.boxDensity = boxDensity
+    this.boxHorizontalAlignment = boxHorizontalAlignment
+    this.boxVerticalAlignment = boxVerticalAlignment
+    with(boxDensity) {
+      boxMarginStart = boxMargin.start.toPxInt()
+      boxMarginEnd = boxMargin.end.toPxInt()
+      boxMarginTop = boxMargin.top.toPxInt()
+      boxMarginBottom = boxMargin.bottom.toPxInt()
+    }
+    this.frameWidth = (boxWidth - boxMarginStart - boxMarginEnd).coerceAtLeast(0)
+    this.boxWidthUnspecified = boxWidthUnspecified
+    this.frameHeight = (boxHeight - boxMarginTop - boxMarginBottom).coerceAtLeast(0)
+    this.boxHeightUnspecified = boxHeightUnspecified
+  }
+
+  /** Measure [widget]. Always call [box] first. */
+  fun measure(widget: Widget<View>) {
+    this.horizontalAlignment = boxHorizontalAlignment
+    this.verticalAlignment = boxVerticalAlignment
+    this.marginStart = 0
+    this.marginEnd = 0
+    this.marginTop = 0
+    this.marginBottom = 0
+    this.requestedWidth = -1
+    this.requestedHeight = -1
+
+    with(boxDensity) {
+      widget.modifier.forEachScoped { childModifier ->
+        when (childModifier) {
+          is HorizontalAlignment -> horizontalAlignment = childModifier.alignment
+          is VerticalAlignment -> verticalAlignment = childModifier.alignment
+          is Width -> requestedWidth = childModifier.width.toPxInt()
+          is Height -> requestedHeight = childModifier.height.toPxInt()
+          is Size -> {
+            requestedWidth = childModifier.width.toPxInt()
+            requestedHeight = childModifier.height.toPxInt()
+          }
+
+          is RedwoodMargin -> {
+            marginStart = maxOf(marginStart, childModifier.margin.start.toPxInt())
+            marginEnd = maxOf(marginEnd, childModifier.margin.end.toPxInt())
+            marginTop = maxOf(marginTop, childModifier.margin.top.toPxInt())
+            marginBottom = maxOf(marginBottom, childModifier.margin.bottom.toPxInt())
           }
         }
       }
     }
 
-    width = if (requestedWidth != Int.MIN_VALUE) {
-      requestedWidth
-    } else {
-      horizontalAlignment.toDimension()
+    val availableWidth = (frameWidth - marginWidth).coerceAtLeast(0)
+    val availableHeight = (frameHeight - marginHeight).coerceAtLeast(0)
+
+    val widthMeasureSpec = when {
+      requestedWidth != -1 -> makeMeasureSpec(requestedWidth, MeasureSpec.EXACTLY)
+      horizontalAlignment == CrossAxisAlignment.Stretch ->
+        makeMeasureSpec(availableWidth, MeasureSpec.EXACTLY)
+      boxWidthUnspecified -> makeMeasureSpec(availableWidth, MeasureSpec.UNSPECIFIED)
+      else -> makeMeasureSpec(availableWidth, MeasureSpec.AT_MOST)
     }
-    height = if (requestedHeight != Int.MIN_VALUE) {
-      requestedHeight
-    } else {
-      verticalAlignment.toDimension()
+
+    val heightMeasureSpec = when {
+      requestedHeight != -1 -> makeMeasureSpec(requestedHeight, MeasureSpec.EXACTLY)
+      verticalAlignment == CrossAxisAlignment.Stretch ->
+        makeMeasureSpec(availableHeight, MeasureSpec.EXACTLY)
+      boxHeightUnspecified -> makeMeasureSpec(availableHeight, MeasureSpec.UNSPECIFIED)
+      else -> makeMeasureSpec(availableHeight, MeasureSpec.AT_MOST)
     }
-    gravity = toGravity(horizontalAlignment, verticalAlignment)
+
+    widget.value.measure(widthMeasureSpec, heightMeasureSpec)
+
+    width = when {
+      requestedWidth != -1 -> requestedWidth
+      horizontalAlignment == CrossAxisAlignment.Stretch -> availableWidth
+      else -> widget.value.measuredWidth
+    }
+
+    height = when {
+      requestedHeight != -1 -> requestedHeight
+      verticalAlignment == CrossAxisAlignment.Stretch -> availableHeight
+      else -> widget.value.measuredHeight
+    }
   }
 
-  private fun toGravity(
-    horizontalAlignment: CrossAxisAlignment,
-    verticalAlignment: CrossAxisAlignment,
-  ): Int {
-    val horizontalGravity = when (horizontalAlignment) {
-      CrossAxisAlignment.Start -> Gravity.START
-      CrossAxisAlignment.Center -> Gravity.CENTER_HORIZONTAL
-      CrossAxisAlignment.End -> Gravity.END
-      CrossAxisAlignment.Stretch -> Gravity.FILL_HORIZONTAL
-      else -> throw AssertionError()
-    }
-    val verticalGravity = when (verticalAlignment) {
-      CrossAxisAlignment.Start -> Gravity.TOP
-      CrossAxisAlignment.Center -> Gravity.CENTER_VERTICAL
-      CrossAxisAlignment.End -> Gravity.BOTTOM
-      CrossAxisAlignment.Stretch -> Gravity.FILL_VERTICAL
-      else -> throw AssertionError()
-    }
-    return horizontalGravity or verticalGravity
-  }
+  /** Layout [widget]. Always call [measure] first. */
+  fun layout(widget: Widget<View>) {
+    val marginLeft: Int
+    val marginRight: Int
+    val alignRight: Boolean
 
-  private fun CrossAxisAlignment.toDimension(): Int {
-    return when (this) {
-      CrossAxisAlignment.Start -> WRAP_CONTENT
-      CrossAxisAlignment.Center -> WRAP_CONTENT
-      CrossAxisAlignment.End -> WRAP_CONTENT
-      CrossAxisAlignment.Stretch -> MATCH_PARENT
-      else -> throw AssertionError()
+    if (layoutDirection == View.LAYOUT_DIRECTION_RTL) {
+      marginLeft = marginEnd + boxMarginEnd
+      marginRight = marginStart + boxMarginStart
+      alignRight = horizontalAlignment == CrossAxisAlignment.Start
+    } else {
+      marginLeft = marginStart + boxMarginStart
+      marginRight = marginEnd + boxMarginEnd
+      alignRight = horizontalAlignment == CrossAxisAlignment.End
     }
+
+    val left = when {
+      horizontalAlignment == CrossAxisAlignment.Center -> {
+        marginLeft + (frameWidth - width) / 2
+      }
+      alignRight -> {
+        (boxMarginStart + frameWidth + boxMarginEnd) - marginRight - width
+      }
+      else -> marginLeft
+    }
+
+    val top = when (verticalAlignment) {
+      CrossAxisAlignment.Center -> {
+        boxMarginTop + marginTop + (frameHeight - height) / 2
+      }
+      CrossAxisAlignment.End -> {
+        boxMarginTop + frameHeight - marginBottom - height
+      }
+      else -> boxMarginTop + marginTop
+    }
+
+    widget.value.layout(left, top, left + width, top + height)
   }
 }